chabges

+++ src/main/java/com/tradesoft/exchanges/dto/response/clientResponse/SymbolsData.java
@@ -11,21 +11,21 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SymbolsData {
     String base_currency;
-    Integer base_currency_scale;
+    Long base_currency_scale;
     String counter_currency;
-    Integer counter_currency_scale;
-    Integer min_price_increment;
-    Integer min_price_increment_scale;
-    Integer min_order_size;
-    Integer min_order_size_scale;
-    Integer max_order_size;
-    Integer max_order_size_scale;
-    Integer lot_size;
-    Integer lot_size_scale;
+    Long counter_currency_scale;
+    Long min_price_increment;
+    Long min_price_increment_scale;
+    Long min_order_size;
+    Long min_order_size_scale;
+    Long max_order_size;
+    Long max_order_size_scale;
+    Long lot_size;
+    Long lot_size_scale;
     String status;
-    Integer id;
-    Integer auction_price;
-    Integer auction_size;
-    Integer auction_time;
-    Integer imbalance;
+    Long id;
+    Long auction_price;
+    Long auction_size;
+    Long auction_time;
+    Long imbalance;
 }
